third_party/WebKit/LayoutTests/http/tests/credentialmanager/credentialscontainer-store-basics.html

<!DOCTYPE html>
<title>Credential Manager: store() basics.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
function stubResolverChecker(c) {
    assert_equals(c, undefined, "FIXME: We're currently always returning 'undefined'.");
    this.done();
}

function stubRejectionChecker(reason) {
    assert_unreached("store() should not reject, but did: " + reason.name);
}

var local = new PasswordCredential({
    id: 'id',
    password: 'pencil',
    name: 'name',
    iconURL: 'https://example.com/icon.png'
});

var federated = new FederatedCredential({
    id: 'id',
    provider: 'https://federation.test/',
    name: 'name',
    iconURL: 'https://example.test/icon.png'
});

async_test(function () {
    navigator.credentials.store().then(
        this.step_func(function () { assert_unreached("store() should reject."); }),
        this.step_func(function (reason) {
            assert_equals(reason.name, "TypeError");
            this.done();
        }));
}, "Verify the basics of store(): PasswordCredential.");

async_test(function () {
    navigator.credentials.store("not a credential").then(
        this.step_func(function () { assert_unreached("store([string]) should reject."); }),
        this.step_func(function (reason) {
            assert_equals(reason.name, "TypeError");
            this.done();
        }));
}, "Verify the basics of store(): PasswordCredential.");

async_test(function () {
    navigator.credentials.store(local).then(
        this.step_func(stubResolverChecker.bind(this)),
        this.step_func(stubRejectionChecker.bind(this)));
}, "Verify the basics of store(): PasswordCredential.");

async_test(function () {
    navigator.credentials.store(federated).then(
        this.step_func(stubResolverChecker.bind(this)),
        this.step_func(stubRejectionChecker.bind(this)));
}, "Verify the basics of store(): FederatedCredential.");

async_test(function () {
    var federated_insecure = new FederatedCredential({
        id: 'id',
        provider: 'https://federation.test/',
        name: 'name',
        iconURL: 'http://example.test/icon.png'
    });
    navigator.credentials.store(federated_insecure).then(
        this.step_func(function () { assert_unreached("store([string]) should reject."); }),
        this.step_func(function (reason) {
            assert_equals(reason.name, "SecurityError");
            this.done();
        }));
}, "Verify iconURL: FederatedCredential.");

async_test(function () {
    var local_insecure = new PasswordCredential({
        id: 'id',
        password: 'pencil',
        name: 'name',
        iconURL: 'http://example.test/icon.png'
    });
    navigator.credentials.store(local_insecure).then(
        this.step_func(function () { assert_unreached("store([string]) should reject."); }),
        this.step_func(function (reason) {
            assert_equals(reason.name, "SecurityError");
            this.done();
        }));
}, "Verify iconURL: PasswordCredential.");
</script>
